whole and sigle

+++ bicseq2_pipeline.sh
@@ -12,6 +12,8 @@ Usage: $command_name [option] <chromosome> <path/to/the/reference> <path/to/the/
 
 Options:
 	-h | --help: show this message
+	--whole: Normalize and segment the entire chromosome
+	--single: Perform normalization and segmentation for each chromosome
 
 	for bicseq2-norm:
 
@@ -77,22 +79,28 @@ do
 					options_norm+="--bin_only "
 					;;
 				fig_norm=*)
-					options_norm+="--${OPTARG} "
+					fig_norm=`echo $OPTARG | sed -e 's/_norm//g'`
+					echo $fig_norm
+					options_norm+="--$fig_norm "
 					;;
 				title_norm=*)
-					options_norm+="--${OPTARG} "
+					title_norm=`echo $OPTARG | sed -e 's/_norm//g'`
+					options_norm+="--$title_norm "
 					;;
 				tmp_norm=*)
-					options_norm+="--${OPTARG} "
+					tmp_norm=`echo $OPTARG | sed -e 's/_norm//g'`
+					options_norm+="--$tmp_norm "
 					;;
 				lambda=*)
 					options_seg+="--${OPTARG} "
 					;;
 				tmp_seg=*)
-					options_seg+="--${OPTARG} "
+					tmp_seg=`echo $OPTARG | sed -e 's/_seg//g'`
+					options_seg+="--$tmp_seg "
 					;;
 				fig_seg=*)
-					options_seg+="--${OPTARG} "
+					fig_segg=`echo $OPTARG | sed -e 's/_seg//g'`
+					options_seg+="--$fig_seg "
 					;;
 				nrm)
 					options_seg+="--nrm"
@@ -183,21 +191,27 @@ function modified_bicseq2-norm() {
 		do
 			echo -e "${line}\t$reference_dir/${line}.fa\t$mappability_dir/${line}.bedgraph\t$working_dir/seq/${line}.seq\t$working_dir/output_norm/${line}.norm.bin" >> $working_dir/config_norm
 		done < $chromosomes
+		touch $working_dir/output_file_norm
+		NBICseq-norm.pl ${options_norm}$working_dir/config_norm $working_dir/output_file_norm
 	fi
 	if (( $single_flag == 1 )); then
 		mkdir -p $working_dir/output_norm_single
 		mkdir -p $working_dir/config_norm_single
 		while read -r line
 		do
-			touch $working_dir/config_norm_single/config_norm_${line}
-			echo -e "chromName\tfaFile\tMapFile\treadPosFile\tbinFileNorm" >> $working_dir/config_norm
-			echo -e "${line}\t$reference_dir/${line}.fa\t$mappability_dir/${line}.bedgraph\t$working_dir/seq/${line}.seq\t$working_dir/output_norm_single/${line}.norm.bin" >> $working_dir/config_norm_${line}
+			if [ -e $working_dir/config_norm_single/config_norm_${line} ]; then
+				echo konnitiha
+				rm $working_dir/config_norm_single/config_norm_${line}
+			fi
 
+			touch $working_dir/config_norm_single/config_norm_${line}
+			touch $working_dir/output_norm_single/output_file_norm_${line}
+			echo -e "chromName\tfaFile\tMapFile\treadPosFile\tbinFileNorm" >> $working_dir/config_norm_single/config_norm_${line}
+			echo -e "${line}\t$reference_dir/${line}.fa\t$mappability_dir/${line}.bedgraph\t$working_dir/seq/${line}.seq\t$working_dir/output_norm_single/${line}.norm.bin" >> $working_dir/config_norm_single/config_norm_${line}
+			NBICseq-norm.pl ${options_norm}$working_dir/config_norm_single/config_norm_${line} $working_dir/output_norm_single/output_file_norm_${line}
 		done < $chromosomes
 	fi
 	echo 'hello'
-	touch $working_dir/output_file_norm
-	NBICseq-norm.pl ${options_norm}$working_dir/config_norm $working_dir/output_file_norm
 
 }
 
@@ -213,16 +227,34 @@ fi
 
 
 # bicseq2-seg
-if [ -e $output_dir/config_seg ]; then
-	rm $output_dir/config_seg
+if (( $whole_flag == 1 )); then
+	if [ -e $output_dir/config_seg ]; then
+		rm $output_dir/config_seg
+	fi
+
+	touch $output_dir/config_seg
+	echo -e "chromName\tbinFileNorm.Case\tbinFileNorm.Control" >> $output_dir/config_seg
+	while read -r line
+	do
+		echo -e "${line}\t$output_dir/case_norm_dir/output_norm/${line}.norm.bin\t$output_dir/control_norm_dir/output_norm/${line}.norm.bin" >> $output_dir/config_seg
+	done < $chromosomes
+
+	touch $output_dir/output_seg
+	NBICseq-seg.pl ${options_seg}$output_dir/config_seg $output_dir/output_seg
 fi
-
-touch $output_dir/config_seg
-echo -e "chromName\tbinFileNorm.Case\tbinFileNorm.Control" >> $output_dir/config_seg
-while read -r line
-do
-	echo -e "${line}\t$output_dir/case_norm_dir/output_norm/${line}.norm.bin\t$output_dir/control_norm_dir/output_norm/${line}.norm.bin" >> $output_dir/config_seg
-done < $chromosomes
-
-touch $output_dir/output_seg
-NBICseq-seg.pl ${options_seg}$output_dir/config_seg $output_dir/output_seg
+if (( $single_flag == 1 )); then
+	mkdir -p $output_dir/config_seg_single
+	mkdir -p $output_dir/output_seg_single
+	while read -r line
+	do
+		if [ -e $output_dir/config_seg_single/config_seg_${line} ]; then
+			echo 'remove'
+			rm $output_dir/config_seg_single/config_seg_${line}
+		fi
+		touch $output_dir/cofig_seg_single/config_seg_${line}
+		echo -e "chromName\tbinFileNorm.Case\tbinFileNorm.Control" >> $output_dir/config_seg_single/config_seg_${line}
+		echo -e "${line}\t$output_dir/case_norm_dir/output_norm_single/${line}.norm.bin\t$output_dir/control_norm_dir/output_norm_single/${line}.norm.bin" >> $output_dir/config_seg_single/config_seg_${line}
+		touch $output_dir/output_seg_single/output_seg_${line}
+		NBICseq-seg.pl ${options_seg}$output_dir/config_seg_single/config_seg_${line} $output_dir/output_seg_single/output_seg_${line}
+	done < $chromosomes
+fi
